Added user_view, Delete for user profile
