+++ UKS_Git_Site/GitApp/templates/base.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+    {% load static %}
+    <link rel="stylesheet" href="{% static 'style.css' %}" />
+    <title>{{ title }}</title>
+  </head>
+
+  <body>
+    <div class="page">
+      <div class="header">
+        <a href="{% url 'index' %}">Base Page</a>
+        &nbsp; {% if request.user.is_authenticated %}
+        <a href="{% url 'user_profile' %}">Profile</a>
+        </br>
+        <a href="{% url 'logout' %}">Logout</a>
+        {% else %}
+        <a href="{% url 'login' %}">Login</a>
+        {% endif %}
+      </div>
+      <div class="left">
+        {% block links %}{% endblock %}
+      </div>
+      <div class="center">{% block content %}{% endblock %}</div>
+    </div>
+  </body>
+</html>
