+++ UKS_Git_Site/GitApp/templates/index_initial.html
@@ -1,15 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
+  <head>
+    <meta charset="UTF-8" />
     {% load static %}
     <title>Initial index</title>
-</head>
-<body>
+  </head>
+  <body>
     <p>Hello, world. You're at the polls index.</p>
-    <br/><br/>
-    <a href="{% url 'test_redis_page'  %}">Link to cached page</a>
-    <a href="{% url 'login'  %}">Login Page</a>
+    <br /><br />
+    <ul>
+      <li><a href="{% url 'test_redis_page'  %}">Link to cached page</a></li>
+      <li><a href="{% url 'login'  %}">Login Page</a></li>
+      <li><a href="{% url 'logout'  %}">Logout Page</a></li>
+      <li><a href="{% url 'user_profile'  %}">Profile</a></li>
+    </ul>
     <a href="{% url 'test'  %}">Test acess(works for user1)</a>
-</body>
+  </body>
 </html>
