+++ UKS_Git_Site/GitApp/views.py
@@ -3,10 +3,7 @@ from .models import Repository
 
 from django.conf import settings
 from django.core.cache.backends.base import DEFAULT_TIMEOUT
-from django.contrib.auth.decorators import login_required, permission_required
 from django.views.decorators.cache import cache_page
-
-from django.contrib.auth import authenticate, login, logout
 
 
 import redis
@@ -31,25 +28,3 @@ def cached_initial(request):
 def single_repo(request, repository_id):
     repo = get_object_or_404(Repository, id=repository_id)
     return HttpResponse("You're looking at repository %s." % repo.name)
-
-@login_required()
-@permission_required('GitApp.test_access', raise_exception=True)
-def test(request):
-    return render(request,'user_profile.html')
-
-#TODO: razdvojiti views
-def user_login(request):
-    if request.method == 'GET':
-        return render(request, "login.html")
-    if request.method == "POST":
-        username = request.POST['username']
-        password = request.POST['password']
-        user = authenticate(request, username=username, password=password)
-        if user:
-            login(request,user)
-            return redirect("test")
-        else:
-            return redirect('login')
-    else:
-        # neuspesno
-        return redirect('index')
